+++ src/components/Planner/PlannerLayout.module.css
@@ -38,23 +38,24 @@
 
 /* PLANNER TIMELINE */
 #planner-timeline-container {
-  margin-left: 1rem;
   background-color: white;
   width: clamp(250px, 30%, 600px);
+  border-right: 2px solid var(--chakra-colors-gray-300);
 }
 
 .timeline-grid {
-  max-height: 100vh;
+  max-height: 87vh;
   display: grid;
   grid-template-columns: 64px repeat(2, 1fr);
   grid-auto-rows: 1fr;
   overflow-y: auto;
+  margin-left: 1rem;
 }
 
 .grid-hour-container {
   min-height: 128px;
   position: relative;
-  border-top: 2px solid var(--chakra-colors-blackAlpha-400);
+  border-bottom: 2px solid var(--chakra-colors-blackAlpha-400);
 }
 
 .timeline-event-wrapper {
